Implemented ! (factorial) as a supported character.
As ! is the only (so far) supported unary operator that follows its
operand, it is at first treated as an operator and then later as a
normal.
The factorial method therefore evalutes based on it's left subtree's
expression and simply multiplies by the right subtree if it exists.
This allows expression like 5!2 to evaluate to (5!) * 2 and 5!^2 to
evaluate to (5!)^2.
I expected that should any other unary operators that follow their
operand come into support, they'd function the same way as a rule:
evaluate the actual operation based on the left subtree, and multiply
by the right if it exists.